Use pushed color package again
While doing local development on color/ the changes were not visible.
I've not written Go in a long time but this was because I'm working on
this project while it's not in the GOPATH, so it was actually loading in
`$GOPATH/src/github.com/bittersweet/imgfiltrate/color` and not my local
file.